ci(playground): move colocated context tests into __tests__ dirs
The changed-test-gate builds the base branch with only the PR's changed
test files restored. The playground build tsconfig only excluded
src/**/__tests__/**, so colocated *.test.tsx files were type-checked
against the base tree where their imported modules don't exist yet
(TS2307). Move the two context tests into __tests__/ and also exclude
src/**/*.test.* from tsconfig.build.json to prevent recurrence.
